[feat] : 본인의 article이나 comment가 아닐경우 exception걸림

+++ server/src/article/service/article.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, Injectable } from '@nestjs/common';
 import { UpdateArticleDto, CreateArticleDto } from '@/article/dto';
 import { ArticleRepository, TagRepository } from '@/article/repository';
 import { Article } from '@/article/entity';
@@ -22,11 +22,23 @@ export class ArticleService {
     await this.articleRepository.createArticle(userId, tags, dto);
   }
 
-  async updateArticle(id: number, dto: UpdateArticleDto): Promise<void> {
-    await this.articleRepository.updateArticle(id, dto);
+  async updateArticle(
+    userId: number,
+    articleId: number,
+    dto: UpdateArticleDto,
+  ): Promise<void> {
+    const article = await this.articleRepository.findArticleById(articleId);
+    if (article.fk_user_id !== userId) {
+      throw new HttpException('This is not your article.', 401);
+    }
+    await this.articleRepository.updateArticle(articleId, dto);
   }
 
-  async deleteArticle(articleId: number): Promise<void> {
+  async deleteArticle(userId: number, articleId: number): Promise<void> {
+    const article = await this.articleRepository.findArticleById(articleId);
+    if (article.fk_user_id !== userId) {
+      throw new HttpException('This is not your article.', 401);
+    }
     await this.articleRepository.deleteArticle(articleId);
   }
 }
